+++ src/client/output/GraphOutput.tsx
@@ -1,3 +1,5 @@
+import * as YAML from 'json-to-pretty-yaml';
+
 import React from 'react';
 import { StandardTableOutput } from '~client/standard_components/app-components';
 
@@ -5,10 +7,20 @@ import { StandardTableOutput } from '~client/standard_components/app-components'
 
 type Props = {
   data: string[];
+  summary: object;
 };
-const GraphOutput = ({ data }: Props) => {
+
+const styles = {
+  pre: {
+    fontWeight: 'bold',
+  },
+};
+const GraphOutput = ({ data, summary }: Props) => {
+  const output = YAML.stringify(summary);
+
   return (
     <div>
+      <pre style={styles.pre}>{output}</pre>
       {!!data ? <StandardTableOutput data={{ rows: data }} tableId="graphOutput" /> : <h2>No Output to display</h2>}
     </div>
   );